requirements.txt
autopep8==2.3.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
feffery-antd-components==0.3.6
feffery-utils-components==0.2.0rc25
flask==3.0.3
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
packaging==24.1
plotly==5.24.1
pycodestyle==2.12.1
python-dotenv==1.0.1
requests==2.32.3
retrying==1.3.4
six==1.16.0
SQLAlchemy==2.0.35
tenacity==9.0.0
tomli==2.0.1
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.4
zipp==3.20.2
